#2469 Review suggestions: don't use MessageHelper in DownstreamMessage
