+++ core/src/main/java/org/eclipse/hono/util/TimeUntilDisconnectNotification.java
@@ -14,6 +14,7 @@
 package org.eclipse.hono.util;
 
 import java.time.Instant;
+import java.util.Objects;
 import java.util.Optional;
 
 import org.apache.qpid.proton.message.Message;
@@ -146,7 +147,7 @@ public final class TimeUntilDisconnectNotification {
 
         if (ttd == null) {
             return Optional.empty();
-        } else if (ttd == 0 || MessageHelper.isDeviceCurrentlyConnected(msg)) {
+        } else if (ttd == 0 || isDeviceCurrentlyConnected(msg)) {
             final String tenantId = MessageHelper.getTenantIdAnnotation(msg);
             final String deviceId = MessageHelper.getDeviceId(msg);
 
@@ -183,5 +184,55 @@ public final class TimeUntilDisconnectNotification {
             final long milliseconds = getReadyUntil().minusMillis(Instant.now().toEpochMilli()).toEpochMilli();
             return (milliseconds > 0 ? milliseconds : 0);
         }
+    }
+
+    /**
+     * Checks if a device is currently connected to a protocol adapter.
+     * <p>
+     * If this method returns {@code true} an attempt could be made to send a command to the device.
+     * <p>
+     * This method uses the message's creation time and TTD value to determine the point in time
+     * until which the device will remain connected.
+     *
+     * @param ttd The TTD value.
+     * @param creationTime The creation time of the message. If {@code null} the device is considered as disconnected.
+     * @return {@code true} if the TTD value contained in the message indicates that the device will
+     *         stay connected for some additional time.
+     */
+    public static boolean isDeviceCurrentlyConnected(final Integer ttd, final Long creationTime) {
+
+        return Optional.ofNullable(ttd).map(ttdValue -> {
+            if (ttdValue == MessageHelper.TTD_VALUE_UNLIMITED) {
+                return true;
+            } else if (ttdValue == 0) {
+                return false;
+            } else {
+                if (creationTime == null) {
+                    return false;
+                }
+
+                final Instant creationTimeInstant = Instant.ofEpochMilli(creationTime);
+                return Instant.now().isBefore(creationTimeInstant.plusSeconds(ttdValue));
+            }
+        }).orElse(false);
+    }
+
+    /**
+     * Checks if a device is currently connected to a protocol adapter.
+     * <p>
+     * If this method returns {@code true} an attempt could be made to send a command to the device.
+     * <p>
+     * This method uses the message's creation time and TTD value to determine the point in time
+     * until which the device will remain connected.
+     *
+     * @param msg The message that is checked for a TTD value.
+     * @return {@code true} if the TTD value contained in the message indicates that the device will
+     *         stay connected for some additional time.
+     * @throws NullPointerException If msg is {@code null}.
+     */
+    public static boolean isDeviceCurrentlyConnected(final Message msg) {
+        Objects.requireNonNull(msg);
+
+        return isDeviceCurrentlyConnected(MessageHelper.getTimeUntilDisconnect(msg), msg.getCreationTime());
     }
 }
